requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
scikit-learn>=1.3.0

# Multi-objective optimization
pymoo>=0.6.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
